feat(query): add native SNII operators and clang benchmark build
Add native query operators for term, boolean, prefix, wildcard, regexp, phrase, and phrase-prefix queries with query profiling and focused boundary/error tests. Extend the benchmark surface so SNII and CLucene results are compared against an oracle, including OSS/sharded execution support and fairer metric reporting.

Align the standalone SNII CMake flags with the Doris BE Clang profile, mark third-party includes as SYSTEM, and centralize checked narrowing at real API boundaries instead of scattering casts through query code.

Tests: cmake --build build-clang-s3 --target snii_core_tests snii_bench -j 16; ./build-clang-s3/core/tests/snii_core_tests; ./build-clang-s3/bench/snii_bench --docs 2000 --vocab 500 --doclen 16 --seed 7; git diff --check.

+++ core/include/snii/writer/compact_posting_pool.h
@@ -128,7 +128,7 @@ class CompactPostingPool {
   //   of (budget exhausted, chain tail reached).
   class Cursor {
    public:
-    Cursor(const CompactPostingPool* pool, uint32_t head, uint32_t budget);
+    Cursor(const CompactPostingPool* pool, uint32_t head, uint64_t budget);
 
     // True while the cursor can still yield a REAL payload byte: the budget is not
     // spent AND the cursor has not reached the chain tail. It peeks the tail forward
@@ -144,14 +144,14 @@ class CompactPostingPool {
     uint32_t cur_;        // absolute read cursor
     uint32_t slice_end_;  // one-past-last payload byte of the current slice
     uint32_t level_;      // current slice level
-    uint32_t budget_;     // remaining byte budget (upper bound on bytes to yield)
+    uint64_t budget_;     // remaining byte budget (upper bound on bytes to yield)
   };
 
   // Builds a cursor over the chain at `head`. `budget` is an UPPER BOUND on bytes to
   // read (see Cursor's contract): the exact payload length or anything larger. The
   // production caller passes the write-head offset, which always bounds the payload
   // from above; the cursor self-terminates at the chain tail regardless.
-  Cursor cursor(uint32_t head, uint32_t budget) const {
+  Cursor cursor(uint32_t head, uint64_t budget) const {
     return Cursor(this, head, budget);
   }
